Refactored all unit tests
