+++ operator/src/operations/conjugate_gradients.rs
@@ -125,48 +125,3 @@ impl<'a, Space: InnerProductSpace, Op: AsApply<Domain = Space, Range = Space>>
         (self.x, rel_res)
     }
 }
-
-#[cfg(test)]
-mod test {
-
-    use crate::interface::{
-        array_vector_space::ArrayVectorSpace, dense_matrix_operator::DenseMatrixOperator,
-    };
-
-    use super::*;
-    use rand::prelude::*;
-    use rlst_dense::rlst_dynamic_array2;
-
-    #[test]
-    fn test_cg() {
-        let dim = 10;
-        let tol = 1E-5;
-
-        let space = ArrayVectorSpace::<f64>::new(dim);
-        let mut residuals = Vec::<f64>::new();
-
-        let mut rng = rand::thread_rng();
-
-        let mut mat = rlst_dynamic_array2!(f64, [dim, dim]);
-
-        for index in 0..dim {
-            mat[[index, index]] = rng.gen_range(1.0..=2.0);
-        }
-
-        let op = DenseMatrixOperator::new(mat.view(), &space, &space);
-
-        let mut rhs = space.zero();
-        rhs.view_mut().fill_from_equally_distributed(&mut rng);
-
-        let cg = (CgIteration::new(&op, &rhs))
-            .set_callable(|_, res| {
-                let res_norm = space.norm(res);
-                residuals.push(res_norm);
-            })
-            .set_tol(tol)
-            .print_debug();
-
-        let (_sol, res) = cg.run();
-        assert!(res < tol);
-    }
-}
