+++ operator/src/operator.rs
@@ -262,46 +262,3 @@ impl<Space: LinearSpace, Op1: AsApply<Range = Space>, Op2: AsApply<Domain = Spac
         self.op2.apply_extended(alpha, &self.op1.apply(x), beta, y)
     }
 }
-
-#[cfg(test)]
-mod test {
-
-    use crate::{
-        interface::{
-            array_vector_space::ArrayVectorSpace, dense_matrix_operator::DenseMatrixOperator,
-        },
-        AsApply, Element, LinearSpace, OperatorBase,
-    };
-    use rlst_dense::{assert_array_relative_eq, rlst_dynamic_array2, traits::*};
-
-    #[test]
-    fn test_operator_algebra() {
-        let mut mat1 = rlst_dynamic_array2!(f64, [4, 3]);
-        let mut mat2 = rlst_dynamic_array2!(f64, [4, 3]);
-
-        let domain = ArrayVectorSpace::new(3);
-        let range = ArrayVectorSpace::new(4);
-
-        mat1.fill_from_seed_equally_distributed(0);
-        mat2.fill_from_seed_equally_distributed(1);
-
-        let op1 = DenseMatrixOperator::new(mat1, &domain, &range);
-        let op2 = DenseMatrixOperator::new(mat2, &domain, &range);
-
-        let mut x = domain.zero();
-        let mut y = range.zero();
-        let mut y_expected = range.zero();
-        x.view_mut().fill_from_seed_equally_distributed(2);
-        y.view_mut().fill_from_seed_equally_distributed(3);
-        y_expected.view_mut().fill_from(y.view());
-
-        op2.apply_extended(2.0, &x, 3.5, &mut y_expected).unwrap();
-        op1.apply_extended(10.0, &x, 1.0, &mut y_expected).unwrap();
-
-        let sum = op1.scale(5.0).sum(op2.as_ref_obj());
-
-        sum.apply_extended(2.0, &x, 3.5, &mut y).unwrap();
-
-        assert_array_relative_eq!(y.view(), y_expected.view(), 1E-12);
-    }
-}
